Add files via upload
Hey guys!
This is a simple python code that takes a csv file as input and creates another csv file and a log file(i.e a text file) with only the required columns we want. An additional step in the end of the code is the required columns can be added to a database as well.
In order to run this code on your pc do this:
1.If you want to insert the data into the database then create one and input the data in the code.(optional)
2.Create a csv file or use on you already have as the parent csv file and provide the location in the code .
3.List the required columns only in the code wherever necessary.
4.Those columns will get inserted to the new csv file and the log file.
Thanks!